pyproject.toml
[project]
name = "my-langchain-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chardet>=5.2.0,<6",
    "chromadb>=1.5.8",
    "dashscope>=1.25.17",
    "ddgs>=9.14.1",
    "duckduckgo-search>=8.1.1",
    "gradio>=5.0.0",
    "langchain>=1.2.15",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.3.0",
    "langchain-deepseek>=1.0.1",
    "langchainhub>=0.1.21",
    "langgraph>=1.1.9",
    "langgraph-checkpoint-sqlite>=2.0.0",
    "markdown>=3.7",
    "openai>=2.32.0",
    "pypdf>=6.10.2",
    "python-dotenv>=1.2.2",
    "unstructured>=0.16.0",
]

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true
